requirements.txt
pytest == 5.0.1
pyyaml == 5.1.2
requests == 2.22.0
PyHamcrest == 1.9.0
allure-pytest == 2.7.1
schematics == 2.1.0
